+++ scriptworker/config.py
@@ -14,9 +14,9 @@ import logging
 import os
 import re
 import sys
+from yaml import safe_load
 
 from scriptworker.constants import DEFAULT_CONFIG
-from scriptworker.client import validate_json_schema
 from scriptworker.context import Context
 from scriptworker.log import update_logging_config
 
@@ -120,26 +120,15 @@ def _is_id_valid(id_string):
     return _GENERIC_ID_REGEX.match(id_string) is not None
 
 
-# _get_json_from_mandatory_file {{{1
-def _get_json_from_mandatory_file(path, function_name):
-    if not os.path.exists(path):
-        print("{} doesn't exist! Exiting {}()...".format(path, function_name),
-              file=sys.stderr)
-        print("Exiting...", file=sys.stderr)
-        sys.exit(1)
-    with open(path, "r", encoding="utf-8") as fh:
-        return json.load(fh)
-
-
 # create_config {{{1
-def create_config(config_path="scriptworker.json"):
+def create_config(config_path="scriptworker.yaml"):
     """Create a config from DEFAULT_CONFIG, arguments, and config file.
 
     Then validate it and freeze it.
 
     Args:
         config_path (str, optional): the path to the config file.  Defaults to
-            "scriptworker.json"
+            "scriptworker.yaml"
 
     Returns:
         tuple: (config frozendict, credentials dict)
@@ -147,7 +136,11 @@ def create_config(config_path="scriptworker.json"):
     Raises:
         SystemExit: on failure
     """
-    secrets = _get_json_from_mandatory_file(config_path, "create_config")
+    if not os.path.exists(config_path):
+        print("{} doesn't exist! Exiting...".format(config_path), file=sys.stderr)
+        sys.exit(1)
+    with open(config_path, "r", encoding="utf-8") as fh:
+        secrets = safe_load(fh)
     config = dict(deepcopy(DEFAULT_CONFIG))
     if not secrets.get("credentials"):
         secrets['credentials'] = read_worker_creds()
@@ -162,31 +155,6 @@ def create_config(config_path="scriptworker.json"):
     del(config['credentials'])
     config = frozendict(config)
     return config, credentials
-
-
-# create_cot_config {{{1
-def create_cot_config(context, cot_config_path=None):
-    """Create a Chain of Trust config from context.config['cot_config'] file.
-
-    Then validate it via the schema file, and freeze it.
-
-    Args:
-        context (scriptworker.context.Context): the scriptworker context.
-
-    Returns:
-        frozendict: the Chain of Trust config.
-
-    Raises:
-        SystemExit: on failure
-    """
-    cot_config_path = cot_config_path or context.config['cot_config_path']
-    cot_config = _get_json_from_mandatory_file(cot_config_path, "create_cot_config")
-    with open(context.config['cot_config_schema_path'], "r") as fh:
-        schema = json.load(fh)
-    validate_json_schema(cot_config, schema, name="cot_config")
-    freeze_values(cot_config)
-    cot_config = frozendict(cot_config)
-    return cot_config
 
 
 # get_context_from_cmdln {{{1
@@ -206,15 +174,10 @@ def get_context_from_cmdln(args, desc="Run scriptworker"):
     context = Context()
     parser = argparse.ArgumentParser(description=desc)
     parser.add_argument(
-        "config_path", type=str, nargs="?", default="scriptworker.json",
+        "config_path", type=str, nargs="?", default="scriptworker.yaml",
         help="the path to the config file"
-    )
-    parser.add_argument(
-        "cot_config_path", type=str, nargs="?",
-        help="the path to the chain of trust config file"
     )
     parsed_args = parser.parse_args(args)
     context.config, credentials = create_config(config_path=parsed_args.config_path)
     update_logging_config(context)
-    context.cot_config = create_cot_config(context, cot_config_path=parsed_args.cot_config_path)
     return context, credentials